FAT: Fix warning
fs/fat/dir.c:43: warning: 'de' may be used uninitialized in this function

The complexity of the code flow makes this seem possible.  Initialize
the value to NULL to eliminate the compiler warning.  This only masks
the warning, since if the value were unused, there would still be a
NULL pointer.

Change-Id: I9fc36abace09409853b63e0997328b75ce703769
Signed-off-by: David Brown <davidb@codeaurora.org>
Signed-off-by: Jcfunk <jncorrea1980@gmail.com>